remove unnecessary jetty abstractions.
* remove JettyHttpUtils, in favor of static methods in Jetty12HttpUtils.
* merge JettyHttpServer and Jetty11Utils into Jetty12HttpServer.
