+++ wiremock-jetty-12/src/main/java/com/github/tomakehurst/wiremock/jetty12/JettySettings.java
@@ -17,10 +17,7 @@ package com.github.tomakehurst.wiremock.jetty12;
 
 import java.util.Optional;
 
-/**
- * Exposed Jetty tuning options. See: <a
- * href="https://www.eclipse.org/jetty/javadoc/jetty-11/org/eclipse/jetty/server/AbstractConnector.html">AbstractConnector</a>
- */
+/** Exposed Jetty tuning options. */
 public class JettySettings {
   private final Integer acceptors;
   private final Integer acceptQueueSize;
